Check recordingPath is not undefined before removeSync

+++ src/index.js
@@ -120,7 +120,9 @@ export default class Video extends WdioReporter {
    * Remove empty directories
    */
   onTestSkip () {
-    fs.removeSync(this.recordingPath);
+    if(this.recordingPath !== undefined) {
+      fs.removeSync(this.recordingPath);
+    }
   }
 
   /**
